requirements.txt
pandas==2.2.2
numpy==2.0.1
yfinance==0.2.43
ta==0.11.0
pytz==2024.1
python-dateutil==2.9.0.post0
requests==2.32.3
streamlit==1.36.0
